src/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-jose==3.3.0
cryptography==3.4.8
python-multipart==0.0.5
passlib==1.7.4
bcrypt==3.2.0
kubernetes==18.20.0
